(WIP) submit button updates profile. Need to look into warning
